Change form request auth to middleware auth

+++ app/Repositories/ProfileRepository.php
@@ -21,7 +21,8 @@ class ProfileRepository
     public function saveProfile($request, $address) {
         $data = $request->only('skills','bio','telegram_username',
             'discord_username','twitter_username','github_username','medium_username','website');
-        $profile = $request->profile;
+        $profile =  $this->model->byAddress($address)->first();
+
         if(!$profile) {
             $data['address'] = strtolower($address);
             $profile = $this->model->create($data);
@@ -35,52 +36,57 @@ class ProfileRepository
         return $profile;
     }
 
-    public function uploadProfileAvatar($request) {
-        $file = $request->file('file');
-        $image = Image::make($file);
-        $height = $image->height();
-        $width = $image->width();
+    public function uploadProfileAvatar($request, $address) {
+        $profile =  $this->model->byAddress($address)->first();
+        if($profile) {
+            $file = $request->file('file');
+            $image = Image::make($file);
+            $height = $image->height();
+            $width = $image->width();
 
-        if($width> 240) {
-            $height = $height * 240/$height;
-            $width = $width * 240/$width;
-        }
-
-        $resized = $image
-            ->resize($width, $height, function ($constraint) { $constraint->aspectRatio(); } )
-            ->encode($file->getCLientOriginalExtension(),80);
-        $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)).'_'.time().'.'.$file->getCLientOriginalExtension();
-        $ret = Storage::put($new_file_name, $resized);
-        if($ret) {
-            $profile = $request->profile;
-            if($profile->avatar && Storage::exists($profile->avatar)) {
-                Storage::delete($profile->avatar);
+            if($width> 240) {
+                $height = $height * 240/$height;
+                $width = $width * 240/$width;
             }
 
-            $profile->avatar = $new_file_name;
-            $profile->save();
-            return $profile;
+            $resized = $image
+                ->resize($width, $height, function ($constraint) { $constraint->aspectRatio(); } )
+                ->encode($file->getCLientOriginalExtension(),80);
+            $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)).'_'.time().'.'.$file->getCLientOriginalExtension();
+            $ret = Storage::put($new_file_name, $resized);
+            if($ret) {
+                if($profile->avatar && Storage::exists($profile->avatar)) {
+                    Storage::delete($profile->avatar);
+                }
+
+                $profile->avatar = $new_file_name;
+                $profile->save();
+                return $profile;
+            }
         }
+
         return null;
     }
 
-    public function uploadProfileBackground($request) {
+    public function uploadProfileBackground($request, $address) {
+        $profile =  $this->model->byAddress($address)->first();
+        if($profile) {
+            $file = $request->file('file');
+            $resized = Image::make($request->file('file'))
+                ->encode($file->getCLientOriginalExtension(),80);
+            $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)).'_'.time().'.'.$file->getCLientOriginalExtension();
+            $ret = Storage::put($new_file_name, $resized);
+            if($ret) {
+                if($profile->background && Storage::exists($profile->background)) {
+                    Storage::delete($profile->background);
+                }
 
-        $file = $request->file('file');
-        $resized = Image::make($request->file('file'))
-            ->encode($file->getCLientOriginalExtension(),80);
-        $new_file_name = Str::slug(pathinfo(basename($file->getClientOriginalName()), PATHINFO_FILENAME)).'_'.time().'.'.$file->getCLientOriginalExtension();
-        $ret = Storage::put($new_file_name, $resized);
-        if($ret) {
-            $profile = $request->profile;
-            if($profile->background && Storage::exists($profile->background)) {
-                Storage::delete($profile->background);
+                $profile->background = $new_file_name;
+                $profile->save();
+                return $profile;
             }
-
-            $profile->background = $new_file_name;
-            $profile->save();
-            return $profile;
         }
+
         return null;
     }
 }
